+++ src/pages/EndUser/Home.js
@@ -1,16 +1,15 @@
 import React from 'react';
-import {Jumbotron} from '../../components/index/jumbotron.jsx';
 import { Featurerette } from '../../components/index/featurerette.jsx';
-import { Features} from '.../../components/index/Features.jsx';
+import {Features} from '../../components/index/Features';
 import {Header} from '../../components/index/Header';
 import HeroHome from '../../components/index/HeroHome';
 import Footer from '../../components/index/Footer';
 
 export function Home() {
-  let body = "Infohorus offers cybersecurity services that combine diverse expertise in the areas of cybersecurity, defensive social engineering, cyber negotiations, intelligence, and other specialized operations to enhance critical infrastructure cyber resilience.";
+  let body = "Infohorus sssoffers cybersecurity services that combine diverse expertise in the areas of cybersecurity, defensive social engineering, cyber negotiations, intelligence, and other specialized operations to enhance critical infrastructure cyber resilience.";
   return (
     <>
-    <Header/>
+    <Header/>  
     <HeroHome/>
     <Features/>
     <Footer/>
